# perf: Calculate length of name column in perf tests